+++ src/refactoring/components/admin/NewProduct.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent } from 'react';
 import { useNewProduct } from '../../hooks/admin/useNewProduct';
-import { createProductWithId } from '../../services/admin';
+import { createProductWithId } from '../../services';
 import { useProductStore } from '../../stores/useProductStore';
 
 export const NewProduct = () => {
