+++ src/refactoring/services/index.ts
@@ -1,3 +1,5 @@
 export * from './admin';
 export * from './cart';
 export * from './coupon';
+export * from './discount';
+export * from './product';
